pyproject.toml
[project]
name = "light-data-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.3",
    "flask-script>=2.0.6",
    "python-dotenv>=1.2.2",
]
